Line 18
#export ENRICO_DIR=/Users/davidsanchez/Soft/enrico

Every time we tried to export the path where enrico is found, using export ENRICO_DIR=/Where/I/put/Enrico/enrico/, it seems that enrico-init.sh do not update it because when we tried to source the init file by using source $ENRICO_DIR/enrico-init.sh it says that we don't have permissions  to create anything in ENRICO_DIR=/Users/davidsanchez/Soft/enrico. So we commented the line and startet again from export and it worked fine. We have installed enrico in others computers before (some months ago) and checking the enrico-init.sh in them, we found this exact line is not there.